ADD: unittest for frole.all, drole.dao

+++ src/main/java/com/wiselink/service/FuncRoleService.java
@@ -14,6 +14,7 @@ import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.wiselink.dao.FuncRoleFuncsDAO;
@@ -85,15 +86,20 @@ public class FuncRoleService {
     /**
      * 设置一个功能能角色#code对应的功能和用户。
      * <p>TODO 性能：rose.dao框架似乎不能批量插入、删除
-     * @param code
+     * @param roleCode
      * @param funcsToDel
      * @param funcsToAdd
      * @param usersToDel
      * @param usersToAdd
+     * @return 
      * @throws ServiceException
      */
-    public void updateFuncRole(int roleCode, List<Integer> funcsToDel, List<Integer> funcsToAdd,
+    public FuncRole updateFuncRole(int roleCode, List<Integer> funcsToDel, List<Integer> funcsToAdd,
             List<String> usersToDel, List<String> usersToAdd) throws ServiceException {
+        FuncRole role = getFuncRoleInfo(roleCode);
+        if (role == null) {
+            return null;
+        }
         try {
             for (int funcCode: funcsToDel) {
                 froleFuncsDao.delete(roleCode, funcCode);
@@ -104,28 +110,59 @@ public class FuncRoleService {
             for (String userId: usersToDel) {
                 froleUsersDao.delete(roleCode, userId);
             }
-            for (String userId: usersToDel) {
-                froleUsersDao.addFuncToRole(roleCode, userId);
+            for (String userId: usersToAdd) {
+                froleUsersDao.addUserToRole(roleCode, userId);
             }
-        } catch (SQLException ex) {
+        } catch (Exception ex) {
             throw new ServiceException(ex);
         }
+        return getFuncRoleList(role);
     }
 
     /**
-     * 获取一个角色#roleCode 对应的所有功能信息
-     * @param roleCode
+     * 修改一个功能角色对应的信息
+     * @param code
+     * @param name
+     * @param desc
+     * @param corpId
+     * @param deptId
+     * @return
+     * @throws ServiceException
+     */
+    public FuncRole updateFuncRole(int code, String name, String desc, String corpId, String deptId) throws ServiceException {
+        boolean ok = false;
+        try {
+            ok = froleDao.update(code, name, desc, corpId, deptId);;
+        } catch (Exception ex) { // SQLException, DataAccessException
+            throw new ServiceException(ex);
+        }
+        if (!ok) {
+            throw new ServiceException("update func role failed.");
+        }
+        return getFuncRole(code);
+    }
+
+    public FuncRole getFuncRole(int code) throws ServiceException {
+        FuncRole role = getFuncRoleInfo(code);
+        if (role == null) {
+            return null;
+        }
+        return getFuncRoleList(role);
+    }
+
+    /**
+     * 获取一个角色#roleCode 对应的所有功能和用户信息
+     * @param frole
      * @return
      * @throws ServiceException 
      */
-    public FuncRole getFuncRole(int roleCode) throws ServiceException {
+    public FuncRole getFuncRoleList(FuncRole frole) throws ServiceException {
         try {
-            FuncRoleInfo froleInfo = froleDao.find(roleCode);
-            List<String> userIds = froleUsersDao.getUsers(roleCode);
+            List<String> userIds = froleUsersDao.getUsers(frole.info.code);
             List<UserCard> users = userService.getUsers(userIds);
-            List<Integer> funcCodes = froleFuncsDao.getFuncs(roleCode);
+            List<Integer> funcCodes = froleFuncsDao.getFuncs(frole.info.code);
             List<Func> funcs = getFuncs(funcCodes);
-            FuncRole frole = new FuncRole(froleInfo).setFuncs(funcs).setUsers(users);
+            frole.setFuncs(funcs).setUsers(users);
             LOGGER.debug("get func role: {}", frole);
             return frole;
         } catch (SQLException ex) {
@@ -134,8 +171,28 @@ public class FuncRoleService {
     }
 
     /**
+     * 获取一个功能角色信息
+     * @param roleCode
+     * @return
+     * @throws ServiceException
+     */
+    public FuncRole getFuncRoleInfo(int code) throws ServiceException {
+        try {
+            FuncRoleInfo froleInfo = froleDao.find(code);
+            if (froleInfo == null) {
+                return null;
+            }
+            LOGGER.debug("get func role info: {}", froleInfo);
+            return new FuncRole(froleInfo);
+        } catch (SQLException ex) {
+            throw new ServiceException(ex);
+        }
+    }
+
+    /**
      * 添加一个新的功能角色
      * 
+     * @param code
      * @param name
      * @param desc
      * @param corpId
@@ -144,11 +201,11 @@ public class FuncRoleService {
      * @return
      * @throws ServiceException
      */
-    public FuncRoleInfo newFuncRole(String name, String desc, String corpId, String deptId, String creatorId) throws ServiceException {
+    public FuncRole newFuncRole(String name, String desc, String corpId, String deptId, String creatorId) throws ServiceException {
         boolean ok = false;
         try {
-            ok = froleDao.add(0, name, desc, corpId, deptId, creatorId);
-        } catch (SQLException ex) {
+            ok = froleDao.add(name, desc, corpId, deptId, creatorId);
+        } catch (Exception ex) { // SQLException, DataAccessException
             throw new ServiceException(ex);
         }
         if (!ok) {
@@ -157,7 +214,7 @@ public class FuncRoleService {
         try {
             FuncRoleInfo frole = froleDao.findByName(name);
             LOGGER.debug("add func role success: {}.", frole);
-            return frole;
+            return new FuncRole(frole);
         } catch (SQLException ex) {
             throw new ServiceException(ex);
         }
@@ -171,4 +228,24 @@ public class FuncRoleService {
         return modules.allModules();
     }
 
+    /**
+     * 获取code指定的功能模块
+     * 
+     * @param code
+     * @return
+     */
+    public FuncModule getModule(int code) {
+        return modules.getModule(code);
+    }
+
+    /**
+     * WARNING: only for unittest or debug.
+     * @throws SQLException 
+     * @throws DataAccessException 
+     */
+    public void clearAll() throws DataAccessException, SQLException {
+        froleDao.clear();
+        froleFuncsDao.clear();
+        froleUsersDao.clear();
+    }
 }
